♻️ refactor: Improve code structure, remove and update unnecessary files
