+++ src/main/java/com/jokim/sivillage/api/brand/Brand.java
@@ -3,18 +3,22 @@ package com.jokim.sivillage.api.brand;
 import jakarta.persistence.*;
 import lombok.Getter;
 
-@Entity
+
 @Getter
+@Entity
 public class Brand {
 
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Long brandId;
+    private Long id;
+
+    @Column(nullable = false, length = 36)
+    private String brandCode;
+
     @Column(length = 30)
     private String mainName;
+
     @Column(length = 30)
     private String subName;
-    private String badgeImage;
-
 
 }
