+++ src/main/java/com/jokim/sivillage/api/product/domain/Product.java
@@ -22,18 +22,23 @@ public class Product {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     private Long id;
+
     @ManyToOne
     @JoinColumn(name = "brand_id")
     private Brand brand;
 
     @Column(length = 255, nullable = false)
     private String productName;
+
     @Column(nullable = true)
     private boolean isOnSale;
+
     @Column(nullable = false)
     private String detail;
+
     @Column(nullable = false)
     private Double price;
+
     @OneToMany(mappedBy = "product")
     private List<ProductOption> options;
