--- src/main/java/com/jokim/sivillage/api/product/domain/ProductSalesVolume.java
@@ -1,5 +1,0 @@
-package com.jokim.sivillage.api.product.domain;
-
-public class ProductSalesVolume {
-
-}
