--- src/main/java/com/jokim/sivillage/api/product/domain/ProductStatistics.java
@@ -1,5 +1,0 @@
-package com.jokim.sivillage.api.product.domain;
-
-public class ProductStatistics {
-
-}
