+++ src/test/java/com/jokim/sivillage/SivillageApplicationTests.java
@@ -1,6 +1,5 @@
 package com.jokim.sivillage;
 
-import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
 import org.springframework.boot.test.context.SpringBootTest;
 
@@ -9,12 +8,6 @@ class SivillageApplicationTests {
 
 	@Test
 	void contextLoads() {
-		System.out.println("Hello World");
-	}
-
-	@BeforeEach
-	void SetUp(){
-		System.out.println("Hello World!!");
 	}
 
 }
